requirements.txt
django==2.2
djangorestframework==3.9.2
flake8>=3.6.0,<3.7.0